+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v0.0.4
+
+* fix crash android
+
+## v0.0.3
+
+* add open system settings function :
+
 ## v0.0.2
 
 * Update readme.md